proto_build

Module message_graph

Structs

DescriptorGraph

MessageGraph builds a graph of messages whose edges correspond to nesting. The goal is to recognize when message types are recursively nested, so that fields can be boxed when necessary.

Enum

EnumValue

Field

FileParser 🔒

Message

Method

A service method descriptor.

OneofField

Service

A service descriptor.